pyproject.toml
[tool.poetry]
name = "pomp-ia"
version = "0.1.0"
description = "Pomp-IA: Plataforma de automatización conversacional y lógica personalizada con modelos de lenguaje."
authors = ["Ariel-Hab <ari.habib.arg@gmail.com>"]
packages = [
  { include = "bot", from = "." },
  { include = "actions", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.11"
make = "^0.1.6.post2"
packaging = ">=20.0,<21.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
tomli = "^2.2.1"

[tool.poetry.group.bot.dependencies]
pandas = ">=1.5.3,<2.0.0"
torch = ">=2.7.1,<3.0.0"
sentencepiece = ">=0.2.0,<0.3.0"
sacremoses = ">=0.1.1,<0.2.0"
unidecode = ">=1.4.0,<2.0.0"
rasa-sdk = "^3.6.1"
rasa = "^3.6.21"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"

[tool.poetry.group.actions.dependencies]
rasa-sdk = "^3.6.1"
rasa = "^3.6.21"
transformers = ">=4.31.0,<4.34.0"  # Versión compatible con Rasa
torch = "^2.5.1"
sentencepiece = "^0.2.0"
openai = "2.3.0"
accelerate = "^0.20.0"
httpx = ">=0.24.0,<0.28.0"
httpcore = ">=0.17.0"

# Scripts
[tool.poetry.scripts]
manage = "scripts.manage:main"
docker = "scripts.docker_helper:menu"
chat = "scripts.chat_console:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"